Fix BehaviorSpec and make it green for the weekend
Issue was that we do validate txs against the confirmed utxo, but
produce a 'TxInvalid' referring the seen utxo.

This requires the test now to wait for a 'SnapshotConfirmed' before
re-submitting.

This likely should be covered by unit tests (we don't have any for
NewTx!) and the benchmark should also wait for snapshots before
re-submitting..ideally.