Add scheduled job to check documentation links on website
Checking links on markdown files while generating documentation is
problematic because some of those links could be to generated pages,
or to missing pages because of other reasons. We had the issue several
times with missing haddock documentation. This job checks the actual
links on the deployed website, crawling all URLs starting from the index.