FIX test that became useless
The removed test used to check that the validator would fail for a snapshot
with a negative value.

But due to recent changes, it is now OK to have a negative snapshot
number. Although it should fail if we try to use a negative snaphot
number with a non-intiail snapshot.

So we amend the previous test we checked for 0 value with a generated
negative or 0 value.