FIX Async issue with tracer
The log tracer used to be multi-threaded, using a queue to store
messages and a thread consuming these messages, posting it to
stdout and flushing stdout.

We observe some race condition at the end of the program execution
that could lead to lost logs.

For instance this test run exhibit the problem: https://github.com/input-output-hk/hydra/actions/runs/4056786570/jobs/6981708131

One thread reads the next message to write from the queue and is
interrupted before actually writing it. Then the _flush_ function is
finally called, reading the full messages queue and writing each of
these messages to stdout. Unfortunately, the message taken but not
written yet is lots.

This could be quite of a bummer in case of problem: the hydra-node
could crash and then, looking at the logs, we would see the last logs
but one missing and that could lead to problematic diagnosis.

We decide to remove this multi-threading for now and proceed as follow:
* all logs are written to sdtout
* we trust _the system_ to manage the buffering and flushing of data
* finally, when the process stops, we explicitly flush all the buffered
  messages.