Maximum number of parties supports is 4
The StateSpec tests do not check the execution budgets, but the tx-cost
benchmark does. However, it was not always trying to abort all commits -
this is changed in this commit now.

In the mid-term, we should fix the property in StateSpec to check
budgets as well or change the claim.