Minor rewording to simplify descriptions.
We want to enforce the why for a mutation to relevant and not the what is doing.

So here we are removing details about the 'what' is failing about the tx.

We think that is better to follow a white box approach on mutations,
thus labels and error codes should be enough to describe the what.