Do not generate reference scripts in genTxOut
We do have several requirements on the TxOut we use in our tests with
the Head protocol. Just lately we needed to "scope out" reference script
support for what can be committed to a Head. Similarly byron addresses
are not supported by plutus.

Hence, this commit also adds a propery that checks whether generated
TxOuts from genTxOut fulfill these requirements.