Fix CollectCom to ensure it collects all value
We seemingly have dropped this technique accidentially. Instead of
checking the head output value against all the collected value, we would
check the other inputs & outputs balance. This is more efficient as we
usually have a more constant number of inputs and outputs to pay for
fees / get change in return.