Write a full script name into the envelope
The text envelope of persisted, golden scripts is a good location to
give a full description of what script is actually included. For this,
let's also incorporate the git describe output, as it gives a very good
indication. Ideally, we change them to the actual tag, but even if not
it's a good starting point.