Rollback carries the chainState
Simplifies the HeadLogic: no need to look for the previous
head state just to find the chain state. The chain component
nows about the chain state to rollback to so it just sends
it in the event.

We don't use the ChainSlot in the event so we could remove it.
However I think it makes sense to share this information in the
event so, not sure, but I'll keep it for now with some sort of
separation of concern idea in mind: the chain component emits
the event that the chain has been rolled back to that point and
it's now in this state. The fact that the information is used
by the receiver or not is not is business.