Introducing an experimental ci alternative
This is doing the same job as CI but trying to embrace github action
pre-installed software on Ubunut images and cache mechanism.

At this stage, the idea is to have both CI live together to observe,
after some time, what happens because there are quite subtle cache
interactions in both strategies.

So let's run both, see what happen and if we can learn something from it.