Filter source of haskell.nix project
This should remove some of the redundant rebuilds and "cache misses" due
to files being changed which do not result in a change to the haskell
results.

Not sure if specifying things to exclude is easier than specifying
directories (= packages) to include. However, as the resulting list can
be bigger than strictly needed (at the cost of redundant rebuilds) I
opted for a blacklist over a whitelist.