Keep track of state between updates
There is quite some room to miss one step for another in the way
some tests are written in the HeadLogicSpec. We stumble upon this
issue working on another PR.

Here we are very humble and just pipe the result of one step
application to another. This is very limited and only deals
with updates which return newState but fits with the actual
use case in the tests.