Remove PostTxError constructors related to wallet
We have already ErrCoverFee type to hold all wallet errors.
Instead of mapping those to different constructors in PostTxError
why don't we have only InternalWalletError that would wrap the
ErrCoverFee and also provide fields for utxo and transaction.
Left a TODO to add To/From json for ErrCoverFee so that we can
have the nicer output.