Create a behavior test to assert nodes ignore init txs they are not part of
This was handled before in the chain layer, but now it is expected to be
in the protocol logic.

We also update signature of IgnoredHeadInitializing to report on the
off-chain identities of a seen Head initTx. This should be equivalent
for most users, but allows us to reason about it in the behavior
logic (where there are no on-chain identities / keys)