Check OnChainId of participants in logic layer
This was done by the chain layer before, but needs to be done now by the
logic layer. This is the only usage of on-chain identities in the
off-chain Head protocol, but it still is required to retain semantics
from before.

This made initializing the Environment a bit hacky, as it is very
specific to the Cardano configuration.