Use already running cardano-node in hydra-cluster
At first we tried to read the genesis shelly json files, but we don't
know the location of shelley genesis keys and should also not rely on
knowing that when using an existing cardano-node.

By fetching the genesis parameters we have a chance to compute the right
block time (assuming the node is not in Byron era) and provide a
RunningNode handle without knowing the genesis json paths.

Co-authored-by: Franco Testagrossa <franco.testagrossa@iohk.io>