Wait for needed UTxO some time
If two RollbackAndForward actions are generated next to each other
then submitTx fails. Waiting for a small amount of time to pass so
that the needed utxo is present fixes this problem.

Ideally the fix would be: Correctly implemented rollbacks in the mock
chain but perhaps this is enough for now.