Decrement on-chain: add checkSignatures
Note that this doesn't test thoroughly the signature verification
since we alter the redeemer signatures only, but it is a bit hard to
test this one since validator checks that run before this one are
ensuring valid signature. Next step is to enhance mutation framework
with the option to alter the input datum in such way to trigger the
signature checks.