Learn section: revised copy, adjusted contents (#1454)
This pull request proposes renaming the 'Core concepts' section to
'Learn'. This change will allow us to consolidate more overviews in this
section, such as the protocol overview and the specification, as
suggested by the team in the documentation questionnaire. The 'Learn'
section will cater to general users interested in core concepts, the
Hydra overview, and related information.

---------

Co-authored-by: Franco Testagrossa <franco.testagrossa@iohk.io>
Co-authored-by: Franco Testagrossa <franco.testagrossa@gmail.com>