Incremental decommit spec (#1479)
Only specification changes for #1057 to get them merged first as
described in #1473.

:nail_care: Marked sections red that are about to change with
incremental decommits.

:nail_care: Adds `decrementTx` which removes UTxO from the head state to
protocol overview, on-chain and off-chain sections.

:nail_care: Updates close, contest and fanout steps accordingly.

:nail_care: Uses a "version counting" scheme to distinguish cases where
the same snapshot is used to decrement and close the head.

:nail_care: Removes $\mathcal{T}_{\mathsf{all}}$ book-keeping and makes
snapshots request full transactions. Adds a note that requesting only
transaction ids is possible if all parties keep an index. This
simplifies the specification (but departs from a 1:1 mapping to
implemented protocol logic in `Hydra.HeadLogic`)

:nail_care: (Re-)introduces snapshot objects to group confirmed
variables (e.g. $\bar{s}, \bar{v}$) which are just kept to refer to them
later on in `close`.

---

<!-- Consider each and tick it off one way or the other -->
* [ ] CHANGELOG updated or not needed
* [x] Documentation updated
* [x] Haddocks not needed
* [ ] No new TODOs introduced or explained herafter
   - [x] ~~Figures are not updated~~
- [ ] Some handwavy-ness when dealing with $tx_\omega$ and interface
between off-chain and on-chain in regards with $\mathsf{CloseType}$ is
unsharp

---------

Co-authored-by: Sasha Bogicevic <sasha.bogicevic@iohk.io>