Introduce a DepositActivated and DepositExpired state changes
These are not strictly needed for the logic, but to update the
pendingDeposits in the state. Also they are very helpful to debug
failing tests.

Only the DepositExpired is mapped to a server output becaues activation
is more of an internal process.

The pattern of determining newActive/newExpired and wanting to determine
the next active given the already updated state is maybe a hint to a
more monadic way of writing the logic functions?