Fix approving of snapshot with deposits with deadline too soon
This is changing the ReqSn message sent between hydra nodes to only
include a deposit transaction id to be considered. This makes it natural
to only approve snapshots where the deposit txid is known and active.

This way we also handle the problem of requesting snapshots with empty
deposit utxo by design.