Fix deposit approval before deposit period
This handles the created time on Deposit to determine whether it is
still inactive or can be requested / signed upon a snapshot request.

Importantly, the snapshot signing must be able to wait for the deposit
to become active and then submit snapshot acknowledgement. To that end,
the TTL and time between re-enqueuing waits is greatly increased now.