Fix the check on invalid decommit
Before we were sending the invalid decommit to both the websocket and
the HTTP endpoint, but only checking the failure in the case of it going
to the websocket. Let's not worry about over-doing the test here; just
send to the websocket and assume that the /decommit endpoint will just
work the same (we tested it above anyway, for the valid decommit).