Permissive-reading of the state file
This change allows us to read the state file even in the presence of
malformed JSON. This is a kind of "best-effort" hack; we don't ever
_expect_ malformed JSON; but if we find it, the best thing we can try to
do is ignore it and see if things still work correctly. That's what we
do here; when we find an invalid parse we emit a warning.